+++ test/libsolidity/syntaxTests/parsing/modifier_invocation.sol
@@ -1,5 +1,5 @@
 contract c {
-    modifier mod1(uint a) { if (msg.sender == address(a)) _; }
+    modifier mod1(uint a) { if (msg.sender == address(uint160(a))) _; }
     modifier mod2 { if (msg.sender == address(2)) _; }
     function f() public mod1(7) mod2 { }
 }
